core values component created
