+++ src/app/join-us/join-us.scss
@@ -49,180 +49,3 @@
     margin-top: 2.25rem;
   }
 }
-
-.join-us-tab-section {
-  background-color: #F4F8FC;
-  padding: 96px 0;
-  margin-top: 90px;
-}
-
-.join-us-tab {
-
-  &__header-container {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  &__header {
-    border-bottom: 1px solid #BEBEBE;
-    width: 561px;
-  }
-
-  &__button {
-    font-family: inherit;
-    text-transform: none;
-    font-weight: 700;
-    font-size: 36px;
-    padding-bottom: 24px;
-  }
-}
-
-
-.volunteer-tab {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  gap: 2rem;
-  width: 100%;
-  margin-top: 32px;
-
-  &__heading-container {
-    width: 50%;
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-
-  }
-
-  &__heading {
-    font-family: inherit;
-    font-weight: 700;
-    font-size: 44px;
-    line-height: 52.8px;
-    letter-spacing: -0.88px;
-    color: #171717;
-  }
-
-  &__sub-heading {
-    color: #737373;
-    font-family: inherit;
-    font-size: 28px;
-    font-style: normal;
-    font-weight: 500;
-    letter-spacing: -0.56px;
-  }
-
-  &__list-container {
-    width: 50%;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: center;
-    gap: 32px;
-  }
-
-  &__item {
-    display: flex;
-    align-items: flex-start;
-    gap: 20px;
-  }
-
-  &__item-number {
-    padding: 8px 14px;
-    background-color: #B9DDFF;
-    border-radius: 6px;
-    font-size: 20px;
-    font-family: inherit;
-    font-weight: 600;
-    color: #0D3C6B;
-    text-align: center;
-  }
-
-  &__item-text-container {
-    display: flex;
-
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 8px;
-  }
-
-  &__item-heading {
-    color: #171717;
-    font-family: inherit;
-    font-size: 28px;
-    font-weight: 700;
-    letter-spacing: -0.56px;
-  }
-
-  &__item-text {
-    color: var(--Neutral-500, #737373);
-    font-family: inherit;
-    font-size: 20px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: normal;
-  }
-}
-
-.partner-tab {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 48px;
-  margin-top: 32px;
-
-  &__heading-container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 8px;
-  }
-
-  &__heading {
-    color: #424242;
-    text-align: center;
-    font-family: inherit;
-    font-size: 28px;
-    font-weight: 500;
-    line-height: normal;
-    letter-spacing: -0.56px;
-  }
-
-  &__sub-heading {
-    color: #424242;
-    font-family: inherit;
-    font-size: 44px;
-    font-weight: 700;
-    letter-spacing: -0.88px;
-    text-align: center;
-  }
-
-  &__item-container {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-  }
-
-  &__item {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    gap: 20px;
-  }
-
-  &__icon {}
-
-  &__item-title {
-    color: #424242;
-    font-family: inherit;
-    font-size: 24px;
-    font-weight: 700;
-    letter-spacing: -0.48px;
-  }
-}
