+++ src/components/core-values/core-values.tsx
@@ -1,27 +1,167 @@
-import { Box, Typography } from '@mui/material';
+import { Box, Container, Typography } from '@mui/material';
 import Image from 'next/image';
 import React from 'react';
 import './core-values.scss';
 
 export default function CoreValues() {
   return (
-    <Box className="sub-container-2-col">
-      <Box className="box-size-20">
+    <Container className="core-value">
+      <Box className="core-value__text-container">
+        <Typography className="page-heading page-heading--secondary">
+          Core Value
+        </Typography>
+        <Box className="core-value__list">
+          <Box className="core-value__item">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="64"
+              height="65"
+              viewBox="0 0 64 65"
+              fill="none"
+              className="core-value__icon"
+            >
+              <g clip-path="url(#clip0_729_5641)">
+                <path
+                  d="M10.6667 36.3313L17.4 29.598L10.6667 22.8646L3.93333 29.598L10.6667 36.3313ZM46.6667 34.9313L53.3333 24.2646L60 34.9313H46.6667ZM32 32.2646C29.7778 32.2646 27.8889 31.4869 26.3333 29.9313C24.7778 28.3758 24 26.4869 24 24.2646C24 21.998 24.7778 20.098 26.3333 18.5646C27.8889 17.0313 29.7778 16.2646 32 16.2646C34.2667 16.2646 36.1667 17.0313 37.7 18.5646C39.2333 20.098 40 21.998 40 24.2646C40 26.4869 39.2333 28.3758 37.7 29.9313C36.1667 31.4869 34.2667 32.2646 32 32.2646ZM32 21.598C31.2444 21.598 30.6111 21.8535 30.1 22.3646C29.5889 22.8758 29.3333 23.5091 29.3333 24.2646C29.3333 25.0202 29.5889 25.6535 30.1 26.1646C30.6111 26.6758 31.2444 26.9313 32 26.9313C32.7556 26.9313 33.3889 26.6758 33.9 26.1646C34.4111 25.6535 34.6667 25.0202 34.6667 24.2646C34.6667 23.5091 34.4111 22.8758 33.9 22.3646C33.3889 21.8535 32.7556 21.598 32 21.598ZM0 48.2646V44.0646C0 42.1091 0.988889 40.5424 2.96667 39.3646C4.94444 38.1869 7.51111 37.598 10.6667 37.598C11.2444 37.598 11.8 37.6091 12.3333 37.6313C12.8667 37.6535 13.3778 37.7091 13.8667 37.798C13.2444 38.6869 12.7778 39.6424 12.4667 40.6646C12.1556 41.6869 12 42.7758 12 43.9313V48.2646H0ZM16 48.2646V43.9313C16 41.0424 17.4778 38.7091 20.4333 36.9313C23.3889 35.1535 27.2444 34.2646 32 34.2646C36.8 34.2646 40.6667 35.1535 43.6 36.9313C46.5333 38.7091 48 41.0424 48 43.9313V48.2646H16ZM53.3333 37.598C56.5333 37.598 59.1111 38.1869 61.0667 39.3646C63.0222 40.5424 64 42.1091 64 44.0646V48.2646H52V43.9313C52 42.7758 51.8556 41.6869 51.5667 40.6646C51.2778 39.6424 50.8444 38.6869 50.2667 37.798C50.7556 37.7091 51.2556 37.6535 51.7667 37.6313C52.2778 37.6091 52.8 37.598 53.3333 37.598ZM32 39.598C29.4667 39.598 27.2 39.9313 25.2 40.598C23.2 41.2647 22.0222 42.0424 21.6667 42.9313H42.4C42 42.0424 40.8111 41.2647 38.8333 40.598C36.8556 39.9313 34.5778 39.598 32 39.598Z"
+                  fill="#707070"
+                />
+              </g>
+              <defs>
+                <clipPath id="clip0_729_5641">
+                  <rect
+                    width="64"
+                    height="64"
+                    fill="white"
+                    transform="translate(0 0.264648)"
+                  />
+                </clipPath>
+              </defs>
+            </svg>
+            <Box className="core-value__item-text-container">
+              <Typography className="core-value__item-title">
+                Inclusivity
+              </Typography>
+              <Typography className="core-value__item-detail">
+                Embrace diversity to fuel innovation and creativity.
+              </Typography>
+            </Box>
+          </Box>
+          <Box className="core-value__item">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="64"
+              height="65"
+              viewBox="0 0 64 65"
+              fill="none"
+              className="core-value__icon"
+            >
+              <g clip-path="url(#clip0_729_5641)">
+                <path
+                  d="M10.6667 36.3313L17.4 29.598L10.6667 22.8646L3.93333 29.598L10.6667 36.3313ZM46.6667 34.9313L53.3333 24.2646L60 34.9313H46.6667ZM32 32.2646C29.7778 32.2646 27.8889 31.4869 26.3333 29.9313C24.7778 28.3758 24 26.4869 24 24.2646C24 21.998 24.7778 20.098 26.3333 18.5646C27.8889 17.0313 29.7778 16.2646 32 16.2646C34.2667 16.2646 36.1667 17.0313 37.7 18.5646C39.2333 20.098 40 21.998 40 24.2646C40 26.4869 39.2333 28.3758 37.7 29.9313C36.1667 31.4869 34.2667 32.2646 32 32.2646ZM32 21.598C31.2444 21.598 30.6111 21.8535 30.1 22.3646C29.5889 22.8758 29.3333 23.5091 29.3333 24.2646C29.3333 25.0202 29.5889 25.6535 30.1 26.1646C30.6111 26.6758 31.2444 26.9313 32 26.9313C32.7556 26.9313 33.3889 26.6758 33.9 26.1646C34.4111 25.6535 34.6667 25.0202 34.6667 24.2646C34.6667 23.5091 34.4111 22.8758 33.9 22.3646C33.3889 21.8535 32.7556 21.598 32 21.598ZM0 48.2646V44.0646C0 42.1091 0.988889 40.5424 2.96667 39.3646C4.94444 38.1869 7.51111 37.598 10.6667 37.598C11.2444 37.598 11.8 37.6091 12.3333 37.6313C12.8667 37.6535 13.3778 37.7091 13.8667 37.798C13.2444 38.6869 12.7778 39.6424 12.4667 40.6646C12.1556 41.6869 12 42.7758 12 43.9313V48.2646H0ZM16 48.2646V43.9313C16 41.0424 17.4778 38.7091 20.4333 36.9313C23.3889 35.1535 27.2444 34.2646 32 34.2646C36.8 34.2646 40.6667 35.1535 43.6 36.9313C46.5333 38.7091 48 41.0424 48 43.9313V48.2646H16ZM53.3333 37.598C56.5333 37.598 59.1111 38.1869 61.0667 39.3646C63.0222 40.5424 64 42.1091 64 44.0646V48.2646H52V43.9313C52 42.7758 51.8556 41.6869 51.5667 40.6646C51.2778 39.6424 50.8444 38.6869 50.2667 37.798C50.7556 37.7091 51.2556 37.6535 51.7667 37.6313C52.2778 37.6091 52.8 37.598 53.3333 37.598ZM32 39.598C29.4667 39.598 27.2 39.9313 25.2 40.598C23.2 41.2647 22.0222 42.0424 21.6667 42.9313H42.4C42 42.0424 40.8111 41.2647 38.8333 40.598C36.8556 39.9313 34.5778 39.598 32 39.598Z"
+                  fill="#707070"
+                />
+              </g>
+              <defs>
+                <clipPath id="clip0_729_5641">
+                  <rect
+                    width="64"
+                    height="64"
+                    fill="white"
+                    transform="translate(0 0.264648)"
+                  />
+                </clipPath>
+              </defs>
+            </svg>
+            <Box className="core-value__item-text-container">
+              <Typography className="core-value__item-title">
+                Inclusivity
+              </Typography>
+              <Typography className="core-value__item-detail">
+                Embrace diversity to fuel innovation and creativity.
+              </Typography>
+            </Box>
+          </Box>
+          <Box className="core-value__item">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="64"
+              height="65"
+              viewBox="0 0 64 65"
+              fill="none"
+              className="core-value__icon"
+            >
+              <g clip-path="url(#clip0_729_5641)">
+                <path
+                  d="M10.6667 36.3313L17.4 29.598L10.6667 22.8646L3.93333 29.598L10.6667 36.3313ZM46.6667 34.9313L53.3333 24.2646L60 34.9313H46.6667ZM32 32.2646C29.7778 32.2646 27.8889 31.4869 26.3333 29.9313C24.7778 28.3758 24 26.4869 24 24.2646C24 21.998 24.7778 20.098 26.3333 18.5646C27.8889 17.0313 29.7778 16.2646 32 16.2646C34.2667 16.2646 36.1667 17.0313 37.7 18.5646C39.2333 20.098 40 21.998 40 24.2646C40 26.4869 39.2333 28.3758 37.7 29.9313C36.1667 31.4869 34.2667 32.2646 32 32.2646ZM32 21.598C31.2444 21.598 30.6111 21.8535 30.1 22.3646C29.5889 22.8758 29.3333 23.5091 29.3333 24.2646C29.3333 25.0202 29.5889 25.6535 30.1 26.1646C30.6111 26.6758 31.2444 26.9313 32 26.9313C32.7556 26.9313 33.3889 26.6758 33.9 26.1646C34.4111 25.6535 34.6667 25.0202 34.6667 24.2646C34.6667 23.5091 34.4111 22.8758 33.9 22.3646C33.3889 21.8535 32.7556 21.598 32 21.598ZM0 48.2646V44.0646C0 42.1091 0.988889 40.5424 2.96667 39.3646C4.94444 38.1869 7.51111 37.598 10.6667 37.598C11.2444 37.598 11.8 37.6091 12.3333 37.6313C12.8667 37.6535 13.3778 37.7091 13.8667 37.798C13.2444 38.6869 12.7778 39.6424 12.4667 40.6646C12.1556 41.6869 12 42.7758 12 43.9313V48.2646H0ZM16 48.2646V43.9313C16 41.0424 17.4778 38.7091 20.4333 36.9313C23.3889 35.1535 27.2444 34.2646 32 34.2646C36.8 34.2646 40.6667 35.1535 43.6 36.9313C46.5333 38.7091 48 41.0424 48 43.9313V48.2646H16ZM53.3333 37.598C56.5333 37.598 59.1111 38.1869 61.0667 39.3646C63.0222 40.5424 64 42.1091 64 44.0646V48.2646H52V43.9313C52 42.7758 51.8556 41.6869 51.5667 40.6646C51.2778 39.6424 50.8444 38.6869 50.2667 37.798C50.7556 37.7091 51.2556 37.6535 51.7667 37.6313C52.2778 37.6091 52.8 37.598 53.3333 37.598ZM32 39.598C29.4667 39.598 27.2 39.9313 25.2 40.598C23.2 41.2647 22.0222 42.0424 21.6667 42.9313H42.4C42 42.0424 40.8111 41.2647 38.8333 40.598C36.8556 39.9313 34.5778 39.598 32 39.598Z"
+                  fill="#707070"
+                />
+              </g>
+              <defs>
+                <clipPath id="clip0_729_5641">
+                  <rect
+                    width="64"
+                    height="64"
+                    fill="white"
+                    transform="translate(0 0.264648)"
+                  />
+                </clipPath>
+              </defs>
+            </svg>
+            <Box className="core-value__item-text-container">
+              <Typography className="core-value__item-title">
+                Inclusivity
+              </Typography>
+              <Typography className="core-value__item-detail">
+                Embrace diversity to fuel innovation and creativity.
+              </Typography>
+            </Box>
+          </Box>
+          <Box className="core-value__item">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="64"
+              height="65"
+              viewBox="0 0 64 65"
+              fill="none"
+              className="core-value__icon"
+            >
+              <g clip-path="url(#clip0_729_5641)">
+                <path
+                  d="M10.6667 36.3313L17.4 29.598L10.6667 22.8646L3.93333 29.598L10.6667 36.3313ZM46.6667 34.9313L53.3333 24.2646L60 34.9313H46.6667ZM32 32.2646C29.7778 32.2646 27.8889 31.4869 26.3333 29.9313C24.7778 28.3758 24 26.4869 24 24.2646C24 21.998 24.7778 20.098 26.3333 18.5646C27.8889 17.0313 29.7778 16.2646 32 16.2646C34.2667 16.2646 36.1667 17.0313 37.7 18.5646C39.2333 20.098 40 21.998 40 24.2646C40 26.4869 39.2333 28.3758 37.7 29.9313C36.1667 31.4869 34.2667 32.2646 32 32.2646ZM32 21.598C31.2444 21.598 30.6111 21.8535 30.1 22.3646C29.5889 22.8758 29.3333 23.5091 29.3333 24.2646C29.3333 25.0202 29.5889 25.6535 30.1 26.1646C30.6111 26.6758 31.2444 26.9313 32 26.9313C32.7556 26.9313 33.3889 26.6758 33.9 26.1646C34.4111 25.6535 34.6667 25.0202 34.6667 24.2646C34.6667 23.5091 34.4111 22.8758 33.9 22.3646C33.3889 21.8535 32.7556 21.598 32 21.598ZM0 48.2646V44.0646C0 42.1091 0.988889 40.5424 2.96667 39.3646C4.94444 38.1869 7.51111 37.598 10.6667 37.598C11.2444 37.598 11.8 37.6091 12.3333 37.6313C12.8667 37.6535 13.3778 37.7091 13.8667 37.798C13.2444 38.6869 12.7778 39.6424 12.4667 40.6646C12.1556 41.6869 12 42.7758 12 43.9313V48.2646H0ZM16 48.2646V43.9313C16 41.0424 17.4778 38.7091 20.4333 36.9313C23.3889 35.1535 27.2444 34.2646 32 34.2646C36.8 34.2646 40.6667 35.1535 43.6 36.9313C46.5333 38.7091 48 41.0424 48 43.9313V48.2646H16ZM53.3333 37.598C56.5333 37.598 59.1111 38.1869 61.0667 39.3646C63.0222 40.5424 64 42.1091 64 44.0646V48.2646H52V43.9313C52 42.7758 51.8556 41.6869 51.5667 40.6646C51.2778 39.6424 50.8444 38.6869 50.2667 37.798C50.7556 37.7091 51.2556 37.6535 51.7667 37.6313C52.2778 37.6091 52.8 37.598 53.3333 37.598ZM32 39.598C29.4667 39.598 27.2 39.9313 25.2 40.598C23.2 41.2647 22.0222 42.0424 21.6667 42.9313H42.4C42 42.0424 40.8111 41.2647 38.8333 40.598C36.8556 39.9313 34.5778 39.598 32 39.598Z"
+                  fill="#707070"
+                />
+              </g>
+              <defs>
+                <clipPath id="clip0_729_5641">
+                  <rect
+                    width="64"
+                    height="64"
+                    fill="white"
+                    transform="translate(0 0.264648)"
+                  />
+                </clipPath>
+              </defs>
+            </svg>
+            <Box className="core-value__item-text-container">
+              <Typography className="core-value__item-title">
+                Inclusivity
+              </Typography>
+              <Typography className="core-value__item-detail">
+                Embrace diversity to fuel innovation and creativity.
+              </Typography>
+            </Box>
+          </Box>
+        </Box>
+      </Box>
+      <Box className="core-value__image-container">
         <Image
-          src="https://source.unsplash.com/random?wallpapers"
+          src="./assets/core-values.png"
           alt="Description"
-          width={500}
-          height={500}
-          style={{ width: '80%', height: 'auto' }}
+          width={562}
+          height={422}
+          className="core-value__image"
         />
       </Box>
-      <Box className="box-size-80">
-        <Typography variant="h6">Innovation</Typography>
-        <Typography variant="body2" paragraph>
-          We are committed to continuous improvement, embracing new ideas that
-          drive both technological and societal advancement.
-        </Typography>
-      </Box>
-    </Box>
+    </Container>
   );
 }
