+++ src/components/join-us-tab/join-us-tab.scss
@@ -1,0 +1,191 @@
+@use '../../app/globals' as *;
+
+
+.join-us-tab-section {
+  background-color: #F4F8FC;
+  padding: 96px 0;
+  margin-top: 290px;
+  position: relative;
+
+  &::before {
+    position: absolute;
+    top: -200px;
+    content: '';
+    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='1152' height='200' viewBox='0 0 1152 200' fill='none'%3E%3Cpath fill-rule='evenodd' clip-rule='evenodd' d='M0 43.4115L48 59.0704C96 74.7292 192 106.047 288 90.3881C384 74.7292 480 12.0938 576 1.6546C672 -8.78464 768 32.9723 864 48.6311C960 64.29 1056 53.8508 1104 48.6311L1152 43.4115V200H1104C1056 200 960 200 864 200C768 200 672 200 576 200C480 200 384 200 288 200C192 200 96 200 48 200H0V43.4115Z' fill='%23F4F8FC'/%3E%3C/svg%3E");
+    width: 100%;
+    height: 200px;
+    background-size: cover;
+  }
+
+}
+
+.join-us-tab {
+
+  &__header-container {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  &__header {
+    border-bottom: 1px solid #BEBEBE;
+    width: 561px;
+  }
+
+  &__button {
+    font-family: inherit;
+    text-transform: none;
+    font-weight: 700;
+    font-size: 36px;
+    padding-bottom: 24px;
+  }
+}
+
+
+.volunteer-tab {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  gap: 2rem;
+  width: 100%;
+  margin-top: 32px;
+
+  &__heading-container {
+    width: 50%;
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+
+  }
+
+  &__heading {
+    font-family: inherit;
+    font-weight: 700;
+    font-size: 44px;
+    line-height: 52.8px;
+    letter-spacing: -0.88px;
+    color: #171717;
+  }
+
+  &__sub-heading {
+    color: #737373;
+    font-family: inherit;
+    font-size: 28px;
+    font-style: normal;
+    font-weight: 500;
+    letter-spacing: -0.56px;
+  }
+
+  &__list-container {
+    width: 50%;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    justify-content: center;
+    gap: 32px;
+  }
+
+  &__item {
+    display: flex;
+    align-items: flex-start;
+    gap: 20px;
+  }
+
+  &__item-number {
+    padding: 8px 14px;
+    background-color: #B9DDFF;
+    border-radius: 6px;
+    font-size: 20px;
+    font-family: inherit;
+    font-weight: 600;
+    color: #0D3C6B;
+    text-align: center;
+  }
+
+  &__item-text-container {
+    display: flex;
+
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 8px;
+  }
+
+  &__item-heading {
+    color: #171717;
+    font-family: inherit;
+    font-size: 28px;
+    font-weight: 700;
+    letter-spacing: -0.56px;
+  }
+
+  &__item-text {
+    color: var(--Neutral-500, #737373);
+    font-family: inherit;
+    font-size: 20px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: normal;
+  }
+}
+
+.partner-tab {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 48px;
+  margin-top: 32px;
+
+  &__heading-container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 8px;
+  }
+
+  &__heading {
+    color: #424242;
+    text-align: center;
+    font-family: inherit;
+    font-size: 28px;
+    font-weight: 500;
+    line-height: normal;
+    letter-spacing: -0.56px;
+  }
+
+  &__sub-heading {
+    color: #424242;
+    font-family: inherit;
+    font-size: 44px;
+    font-weight: 700;
+    letter-spacing: -0.88px;
+    text-align: center;
+  }
+
+  &__item-container {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 100%;
+  }
+
+  &__item {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 20px;
+  }
+
+  &__icon {}
+
+  &__item-title {
+    color: #424242;
+    font-family: inherit;
+    font-size: 24px;
+    font-weight: 700;
+    letter-spacing: -0.48px;
+  }
+}
